backend/requirements.txt

# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0

# MongoDB
djongo==1.3.6
pymongo==3.12.3

# ML Integration
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.26.0

# Utilities
Pillow==10.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0

# Development
black==23.9.1
flake8==6.1.0
pytest==7.4.2
pytest-django==4.5.2
